Change webpack configuration
As mentioned in the previous commit, the webpack
configuration has been tweaked to have prod and
dev config files, as well as a common file as
their skeleton. The npm scripts have been modified
accordingly to run with the proper config files.
THey have been tested and work. This should be the
final commit. The project is now complete.